Added 4 more scripts
ui.r and server.r work in conjuction to create the interactive plot for EDA.  Logistic Model.R is the answer to the first part of the case study: what are the top 3 factors that contribute to turnover.  Interactive_Plot.rmd is the actual plot itself.  Unfortunately, I can't make an .md or .html for this type of document, so it'll just have to be posted separately.