Finalized changed to Interactive_Plot.rmd
This is the finished product.  It works beautifully.  Simply set the working directory (line 17 in Interactive_Plot.rmd) to one level up from where your data is located (or read in the data directly on line 18) and hit "Run Document" at the top next to the green arrow.  I'm super proud of this and can't wait to present it to the class!